Refactor chatbot utilities
