+++ main.py
@@ -1,47 +1,8 @@
+"""CLI and GUI entry points for the HyperCLOVAX chatbot."""
+
 import argparse
 
-MODEL_NAME = "naver-hyperclovax/HyperCLOVAX-SEED-Vision-Instruct-3B"
-
-
-def load_model():
-
-
-
-    """Load tokenizer, processor, and model in float32 on CPU."""
-
-    import torch
-    from transformers import (
-        AutoModelForCausalLM,
-        AutoTokenizer,
-        AutoProcessor,
-    )
-
-
-    # Both the vision and language components must run in float32
-
-    tokenizer = AutoTokenizer.from_pretrained(MODEL_NAME)
-    processor = AutoProcessor.from_pretrained(MODEL_NAME)
-    model = AutoModelForCausalLM.from_pretrained(
-        MODEL_NAME,
-        torch_dtype=torch.float32,
-        device_map="cpu",
-    )
-    model.eval()
-    return tokenizer, processor, model
-
-
-def generate_reply(tokenizer, processor, model, text, image_path=None):
-
-  
-    from PIL import Image
-    image = Image.open(image_path).convert("RGB") if image_path else None
-    inputs = processor(text=[text], images=[image] if image else None, return_tensors="pt")
-    import torch
-    with torch.no_grad():
-        outputs = model.generate(**inputs)
-
-    reply = tokenizer.decode(outputs[0], skip_special_tokens=True)
-    return reply
+from chatbot import generate_reply, load_model
 
 
 def chat_cli(args):
