+++ project_file.md
@@ -1,0 +1,8 @@
+# Development Notes
+
+This file logs recent changes made to ensure the example chatbot runs.
+
+- Created `chatbot.py` containing model loading and reply generation functions.
+- Updated `main.py` to use the new module and cleaned duplicate code.
+- Improved error handling when required libraries are missing.
+- Adjusted unit tests to import from `chatbot.py`.
