+++ tests/test_image_mock.py
@@ -6,7 +6,7 @@ try:
 except ImportError:  # pragma: no cover - skip if Pillow missing
     PIL = None
 
-from main import generate_reply
+from chatbot import generate_reply
 
 class DummyTokenizer:
     def decode(self, outputs, skip_special_tokens=True):
